C++FinalProject_Scanner
This project is the program like a scanner. The original intention of our idea is to build an 
efficient tool for students to submit their handwriting homework. The main part of our project 
can be divided into three parts; First, we applied the Hough transform to detect the edge of the 
A4 paper in the picture. In addition, we designed a page which enable the client to manually 
select the paper corner to make up the inaccuracy of the first algorithm. Second, we designed 
two algorithms which can enhance the handwriting of the scanned paper. Finally, we can export 
the scanned document into a single pdf form. In addition to the main function of our project, 
we pay a lot effort to make our product more user friendly because our primary goal is to make 
students submit their homework easier.